giỏ hàng

+++ view/cart/view-cart.php
@@ -18,7 +18,7 @@
                         </tr>
                     </thead>
                     <tbody>
-                    <?php
+                        <?php
                         $totalAllCart = 0;
                         $totalPayPriceOriginal = 0;
                         $formatTotalAllCart = 0;
@@ -29,43 +29,48 @@
 
                         foreach ($gio_hang_all as $gio_hang) {
                             extract($gio_hang);
-                            
+
                             // Kiểm tra xem sản phẩm đã được hiển thị hay chưa
-                            if (array_key_exists($ten_san_pham, $displayedProducts)) {
+                            if (array_key_exists($id_chi_tiet_san_pham, $displayedProducts)) {
                                 // Nếu đã tồn tại, tăng số lượng
-                                $displayedProducts[$ten_san_pham]['so_luong'] += $so_luong;
+                                $displayedProducts[$id_chi_tiet_san_pham]['so_luong'] += $so_luong;
                             } else {
                                 // Nếu chưa tồn tại, lưu sản phẩm vào mảng
-                                $displayedProducts[$ten_san_pham] = [
+                                $displayedProducts[$id_chi_tiet_san_pham] = [
+                                    'ten_san_pham' =>$ten_san_pham,
                                     'hinh_anh' => $hinh_anh,
                                     'don_gia' => $don_gia,
                                     'so_luong' => $so_luong,
-                                    'id_gio_hang' => $id_gio_hang
+                                    'id_gio_hang' => $id_gio_hang,
+                                    'do_an_them' => $do_an_them,
+                                    'khau_phan' => $khau_phan,
                                 ];
                             }
                         }
 
                         // Duyệt qua mảng sản phẩm đã được xử lý
-                        foreach ($displayedProducts as $ten_san_pham => $product) {
+                        foreach ($displayedProducts as $id_chi_tiet_san_pham => $product) {
                             $showImage = $imagePath . $product['hinh_anh'];
                             // Format money
                             $formatPrice = formatCurrency($product['don_gia']);
                             $totalPrice = $product['don_gia'] * $product['so_luong'];
                             $formatTotalMoney = formatCurrency($totalPrice);
-                            
+
                             // Tính tổng số tiền
                             $tong += $totalPrice;
                             $formatTotalAllCart = formatCurrency($tong);
 
                             // Remove item cart
-                            $deleteItemCart = '<a href="index.php?act=delete-cart&id-cart=' . $product['id_gio_hang'] . '"><button class="remove-item-cart" value="Xóa khỏi giỏ hàng"><i class="fa-solid fa-trash-can"></i></button></a>';
-                            
+                            $deleteItemCart = '<a onclick="return confirm(\'Bạn có muốn xóa không?\');" href="index.php?act=delete-cart&id-cart=' .$id_chi_tiet_san_pham . '"><button class="remove-item-cart" value="Xóa khỏi giỏ hàng"><i class="fa-solid fa-trash-can"></i></button></a>';
+
                             // Hiển thị sản phẩm
                             echo '<tr>
                                 <td scope="row">
                                     <img src="' . $showImage . '" alt="Ảnh sản phẩm" width="100" height="100">
                                 </td>
-                                <td class="product-name"><strong>' . $ten_san_pham . '</strong></td>
+                                <td class="product-name"><strong>' . $ten_san_pham . '</strong> <br>
+                                <p style="font-size: 10px;">'.$khau_phan.' & '.$do_an_them.' </p>
+                                </td>
                                 <td class="price-product">' . $formatPrice . '</td>
                                 <td class="quantity-product">' . $product['so_luong'] . '</td>
                                 <td>' . $formatTotalMoney . '</td>
@@ -74,6 +79,7 @@
                         }
                         ?>
 
+
                     </tbody>
                 </table>
                 <div class="shopping-cart-wrapper">
@@ -81,7 +87,7 @@
                         <a href="index.php?act=home">Tiếp tục mua hàng</a>
                     </div>
                     <div class="clear-cart-all">
-                        <a href="index.php?act=delete-cart">Xóa giỏ hàng</a>
+                        <a onclick="return confirm('Bạn muốn xóa tất cả giỏ hàng ?')" href="index.php?act=delete-cart">Xóa giỏ hàng</a>
                     </div>
                 </div>
                 <div class="grand-total">
@@ -89,7 +95,7 @@
                         <h4 class="cart-bottom-title">Tổng tiền giỏ hàng</h4>
                     </div>
                     <h4 class="grand-total-title">Tổng cộng: <span><?= $formatTotalAllCart ?></span></h4>
-                    <a href="index.php?act=checkout">Đặt hàng</a> ;
+                    <a href="index.php?act=checkout">Đặt hàng</a> 
                 </div>
             <?php } else { ?>
                 <div class="no-cart"><img
